Make FloatPrecision = float in TypeDefinitions.h compilable without Vectorization
Vectorization support for this needs a rewrite of the AVX functor for floating point precisions.

For the rest, unfortunately, it was necessary to change a lot, and some stuff has to reviewed carefully:
- The functors need a good look, especially for the globals stuff. I don't know where double precision is still wanted.
- Currently, all boundary calculations and box calculations are still done in double precision. This makes a lot of casts necessary. Is this the way to go?
- For some tests, it seems like they used both TypeDefinitions.h and commonTypedefs.h types. One example are the GridCube tests. I was confused with what to use there.
- Probably most critical: In ArrayMath.h, I added a second float type to some methods to allow operations on data with different precisions. The return type currently is always chosen to be the type of the first array parameter. This might be counterintuitive for some applications.

+++ examples/md-flexible/src/TypeDefinitions.h
@@ -12,7 +12,7 @@
 /**
  * Precision used for particle representations. If you want to test other precisions change it here.
  */
-using FloatPrecision = double;
+using FloatPrecision = float;
 
 /**
  * Type of the Particles used in md-flexible.
